pyproject.toml
[project]
name = "sherpa-technical-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "llama-cloud-services>=0.6.12",
    "llama-extract==0.0.5",
    "llama-index>=0.11.23",
    "llama-index-core>=0.11.23",
    "llama-index-readers-file>=0.3.0",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20250327",
    "pi-heif>=0.22.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "unstructured>=0.17.2",
    "unstructured-inference>=0.8.10",
    "unstructured-pytesseract>=0.3.15",
    "weaviate-client>=4.13.2",
]
